requirements.txt
fastapi
python-jose
passlib
pymongo
bcrypt
python-multipart